check_key_duplicates moved to utils/misc.py
