+++ monai/bundle/config_parser.py
@@ -25,6 +25,7 @@ from monai.bundle.reference_resolver import ReferenceResolver
 from monai.bundle.utils import ID_REF_KEY, ID_SEP_KEY, MACRO_KEY
 from monai.config import PathLike
 from monai.utils import ensure_tuple, look_up_option, optional_import
+from monai.utils.misc import check_key_duplicates
 
 if TYPE_CHECKING:
     import yaml
@@ -385,29 +386,6 @@ class ConfigParser:
         else:
             self.ref_resolver.add_item(ConfigItem(config=item_conf, id=id))
 
-    @staticmethod
-    def check_key_duplicates(ordered_pairs: Sequence[tuple[Any, Any]]) -> dict[Any, Any]:
-        """
-        Checks if there is a duplicated key in the sequence of `ordered_pairs`.
-        If there is - it will log a warning or raise ValueError
-        (if configured by environmental var `MONAI_FAIL_ON_DUPLICATE_CONFIG==1`)
-
-        Otherwise, it returns the dict made from this sequence.
-
-        Args:
-            ordered_pairs: sequence of (key, value)
-        """
-        keys = set()
-        for k, _ in ordered_pairs:
-            if k in keys:
-                if os.environ.get("MONAI_FAIL_ON_DUPLICATE_CONFIG", "0") == "1":
-                    raise ValueError(f"Duplicate key: `{k}`")
-                else:
-                    warnings.warn(f"Duplicate key: `{k}`")
-            else:
-                keys.add(k)
-        return dict(ordered_pairs)
-
     @classmethod
     def load_config_file(cls, filepath: PathLike, **kwargs: Any) -> dict:
         """
@@ -425,7 +403,7 @@ class ConfigParser:
             raise ValueError(f'unknown file input: "{filepath}"')
         with open(_filepath) as f:
             if _filepath.lower().endswith(cls.suffixes[0]):
-                return json.load(f, object_pairs_hook=cls.check_key_duplicates, **kwargs)  # type: ignore[no-any-return]
+                return json.load(f, object_pairs_hook=check_key_duplicates, **kwargs)  # type: ignore[no-any-return]
             if _filepath.lower().endswith(cls.suffixes[1:]):
                 return yaml.safe_load(f, **kwargs)  # type: ignore[no-any-return]
             raise ValueError(f"only support JSON or YAML config file so far, got name {_filepath}.")
